+++ NEWS.md
@@ -1,5 +1,7 @@
 # ggh4x (development version)
 
+* Deprecated `ggsubset()`. Use `data = ~ subset(.x, ...)` instead.
+
 # ggh4x 0.2.1
 
 This is a patch release fixing a few bugs and a broken unit test.